DMRG: Fix bug on svd module
* It was using the potentially uninitialized general logger rather than the one for svd

+++ source/math/svd/settings.h
@@ -3,7 +3,7 @@
 namespace svd {
     struct settings {
         std::optional<double> threshold   = std::nullopt;
-        std::optional<long>   switchsize  = std::nullopt;
+        std::optional<size_t> switchsize  = std::nullopt;
         std::optional<size_t> loglevel    = std::nullopt;
         std::optional<bool>   use_bdc     = std::nullopt;
         std::optional<bool>   use_lapacke = std::nullopt;
